+++ ninja-core/src/main/java/ninja/jpa/JpaConstant.java
@@ -1,0 +1,7 @@
+package ninja.jpa;
+
+public interface JpaConstant {
+    
+    String PERSISTENCE_UNIT_NAME = "ninja.jpa.persistence_unit_name";
+
+}