+++ ninja-jpa-demo/src/main/java/db/migration/1__.sql
@@ -1,2 +1,0 @@
-/** Your first migration goes here */
-/** more: http://flywaydb.org/documentation/migration/sql.html */