+++ ninja-jpa-demo/src/main/java/db/migration/V2__.sql
@@ -1,0 +1,1 @@
+-- the second script that will be run by Ninja's migration engine script