+++ ninja-jpa-demo/src/test/java/controllers/ApplicationControllerTest.java
@@ -37,7 +37,7 @@ public class ApplicationControllerTest extends NinjaTest {
         Map<String, String> headers = Maps.newHashMap();
 
         String result = ninjaTestBrowser.makeRequest(getServerAddress()
-                + "/index", headers);
+                + "/", headers);
 
 
         assertTrue(result
@@ -52,12 +52,49 @@ public class ApplicationControllerTest extends NinjaTest {
         formParameters.put("email", "test2@email.com");
 
         ninjaTestBrowser.makePostRequestWithFormParameters(getServerAddress()
-                + "/index", headers, formParameters);
+                + "/", headers, formParameters);
 
         ///////////////////////////////////////////////////////////////////////
         // STEP 2: Check that posting works (and database, too)
         ///////////////////////////////////////////////////////////////////////
-        result = ninjaTestBrowser.makeRequest(getServerAddress() + "/index",
+        result = ninjaTestBrowser.makeRequest(getServerAddress() + "/",
+                headers);
+
+        assertTrue(result.contains("a text"));
+        assertTrue(result.contains("test2@email.com"));
+    }
+    
+    
+    @Test
+    public void testThatPostingWorksTwoTimesAndDatabseIsResetInBetween() {
+
+        ///////////////////////////////////////////////////////////////////////
+        // STEP 1: Check that db is empty
+        ///////////////////////////////////////////////////////////////////////
+        Map<String, String> headers = Maps.newHashMap();
+
+        String result = ninjaTestBrowser.makeRequest(getServerAddress()
+                + "/", headers);
+
+
+        assertTrue(result
+                .contains("No guestbook entries yet... please enter one above..."));
+
+        ///////////////////////////////////////////////////////////////////////
+        // STEP 2: Post a new guestbookentry.
+        ///////////////////////////////////////////////////////////////////////
+        Map<String, String> formParameters = Maps.newHashMap();
+
+        formParameters.put("text", "a text");
+        formParameters.put("email", "test2@email.com");
+
+        ninjaTestBrowser.makePostRequestWithFormParameters(getServerAddress()
+                + "/", headers, formParameters);
+
+        ///////////////////////////////////////////////////////////////////////
+        // STEP 2: Check that posting works (and database, too)
+        ///////////////////////////////////////////////////////////////////////
+        result = ninjaTestBrowser.makeRequest(getServerAddress() + "/",
                 headers);
 
         assertTrue(result.contains("a text"));