+++ ninja-core/src/site/markdown/developer/changelog.md
@@ -1,8 +1,9 @@
-Version XXX
-===========
+Version 1.6.0
+=============
 
  * 2013-07-17 JPA support went into ninja-core (ra)
  * 2013-07-17 Added a JPA demo acting as integration test and an archetype for JPA (ra)
+ * 2013-07-26 Added Flyway integration (database migration tool) (ra)
 
 Version 1.5.1
 =============